+++ src/js/util/player.js
@@ -1,6 +1,6 @@
 import { isTag } from './dom';
 import { once } from './event';
-import { includes, noop } from './lang';
+import { includes } from './lang';
 
 export function play(el) {
     if (isIFrame(el)) {
@@ -8,11 +8,7 @@ export function play(el) {
     }
 
     if (isHTML5(el)) {
-        try {
-            el.play().catch(noop);
-        } catch (e) {
-            // noop
-        }
+        el.play();
     }
 }
 
@@ -64,11 +60,7 @@ async function call(el, cmd) {
 }
 
 function post(el, cmd) {
-    try {
-        el.contentWindow.postMessage(JSON.stringify({ event: 'command', ...cmd }), '*');
-    } catch (e) {
-        // noop
-    }
+    el.contentWindow.postMessage(JSON.stringify({ event: 'command', ...cmd }), '*');
 }
 
 const stateKey = '_ukPlayer';