+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.16.27 (September 28, 2023)
+
+### Fixed
+
+- Fix Height Viewport component not observing viewport changes with `expand: true`
+- Fix passing a `vh` unit to option `scale` relates to element's height in Parallax component 
+
 ## 3.16.26 (August 28, 2023)
 
 ### Fixed